requirements.txt
tensorflow==2.12.0
numpy==1.24.0
matplotlib==3.8.0
opencv-python==4.10.0
Pillow==9.4.0
streamlit==1.26.0